Flowcharts for simulating loud chirp (works!)
Previous attempts to generate ionograms failed because
I forgot to take the complex conjugate before mixing
the template and the target. This has been fixed, however
we also need to load up matlab and get the exact chirp
rate for the long IQ file in order to get sufficient
SNR on the mixed "filter" output.